Move injecting accessToken from IFrameViewer into BlockPreview
This is required for block preview only, but IFrameViewer is common
